Long values.

+++ Smoke/app/src/main/java/org/purple/smoke/NeighborElement.java
@@ -49,7 +49,7 @@ public class NeighborElement
     public String m_uptime = "";
     public byte m_remoteCertificate[] = null;
     public int m_oid = -1;
-    public int m_outboundQueued = 0;
+    public long m_outboundQueued = 0;
 
     public NeighborElement()
     {
